+++ golang/scripts/build-and-run.sh
@@ -20,7 +20,7 @@ custom_params_json='{
 
 bash "${repo_root_dirpath}/.kurtosis/build-and-run-core.sh" \
     "${action}" \
-    "${KURTOSIS_DOCKERHUB_ORG}/kurtosis-go-example" \
+    "${KURTOSIS_DOCKERHUB_ORG}/kurtosis-golang-example" \
     "${lang_root_dirpath}" \
     "${lang_root_dirpath}/testsuite/Dockerfile" \
     "${repo_root_dirpath}/.kurtosis/kurtosis.sh" \