+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # TBD
+### Added
+* Rust example testsuite to the set of testsuites that get checked in the CircleCI config
+
 ### Changed
+* Refactored CircleCI config to support validating multiple testsuites & pushing multiple example testsuite Docker images
+* Modified CircleCI building to skip building a language's testsuite if a) no shared code has changed and b) the language's directory doesn't have any changes
 * Renamed `DockerContainerInitializer.GetFilesToMount` to `GetFilesToGenerate`, and `InitializeMountedFiles` to `InitializeGeneratedFiles`, in both Rust and Go
 
 # 1.9.0