+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# TBD
+
+# 1.10.0
+### Added
+* Rust example testsuite to the set of testsuites that get checked in the CircleCI config
+* A `release.sh` script which will run the process of cutting a new release for this repo (necessary because we have to update the `version` key in Rust's Cargo.toml files)
+
+### Changed
+* Refactored CircleCI config to support validating multiple testsuites & pushing multiple example testsuite Docker images
+* Modified CircleCI building to skip building a language's testsuite if a) no shared code has changed and b) the language's directory doesn't have any changes
+* Renamed `DockerContainerInitializer.GetFilesToMount` to `GetFilesToGenerate`, and `InitializeMountedFiles` to `InitializeGeneratedFiles`, in both Rust and Go
+
 # 1.9.0
 ### Added
 * Rust client library